+++ starters/vue3-apollo-quasar/src/components/NumberCounter/NumberCounter.vue
@@ -44,7 +44,7 @@ export default defineComponent({
 </script>
 
 <script lang="ts" setup>
-import { increment, decrement, reset } from 'src/variables/counts';
+import { increment, decrement, reset } from 'src/globals/counts';
 import { useQuery } from '@vue/apollo-composable';
 import gql from 'graphql-tag';
