+++ starters/vue3-apollo-quasar/src/init/setupGraphQL.ts
@@ -10,7 +10,7 @@ import { setContext } from '@apollo/client/link/context';
 
 import fetch from 'cross-fetch';
 
-import { counts } from '../globalvariables/counts';
+import { counts } from '../globals/counts';
 
 // HTTP connection to the API
 const httpLink = createHttpLink({
